app/web/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
prometheus-client==0.20.0
psycopg2-binary==2.9.9
redis==5.0.7
